requirements.txt
Django==1.8.5
argparse==1.2.1
django-crispy-forms==1.5.2
django-easy-pjax==1.2.0
django-filter==0.11.0
django-grappelli==2.7.2
django-ldapdb==0.4.0
django-material==0.4.1
djangorestframework==3.3.2
funcparserlib==0.3.6
funcsigs==0.4
mock==1.3.0
mockldap==0.2.6
pbr==1.8.1
python-ldap==2.4.22
six==1.10.0
wsgiref==0.1.2
